refactored tub kinematic check into another file
